% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{download_dhs_datasets_by_type}
\alias{download_dhs_datasets_by_type}
\title{Download DHS Datasets for Specified Countries}
\usage{
download_dhs_datasets_by_type(
  country_codes,
  survey_type,
  file_type,
  file_format,
  output_dir,
  clear_cache = TRUE
)
}
\arguments{
\item{country_codes}{A character vector of ISO3 country codes.}

\item{survey_type}{A character string. Type of DHS survey (e.g., "DHS").}

\item{file_type}{A character string. Type of file (e.g., "PR" or "GE").}

\item{file_format}{A character string. Format of the file (e.g., "DT",
"FLAT").}

\item{output_dir}{A character string specifying the directory to save
downloaded datasets.}

\item{clear_cache}{Logical. Whether to clear the cache after downloading.
Default is TRUE.}
}
\value{
Invisibly returns a vector of downloaded dataset filenames.
}
\description{
Downloads DHS datasets of a specified survey type and file type for given
countries.
}
