% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{generate_predictor_data}
\alias{generate_predictor_data}
\title{Generate or Load Cached Predictors Data}
\usage{
generate_predictor_data(
  country_code,
  country_shape,
  pop_raster,
  ur_raster,
  adm2_shape,
  cell_size = 5000,
  ignore_cache = FALSE,
  output_dir = "03_outputs/3a_model_outputs"
)
}
\arguments{
\item{country_code}{A string representing the country code (e.g., "KEN").}

\item{country_shape}{An `sf` object representing the country's administrative
boundaries.}

\item{pop_raster}{A `terra` raster object representing the population raster.}

\item{ur_raster}{A `terra` raster object representing the urban extent
raster.}

\item{adm2_shape}{An `sf` object representing the administrative level 2
boundaries.}

\item{cell_size}{An integer specifying the cell size for the prediction grid
in meters (default is 5000).}

\item{ignore_cache}{A boolean input which is set to determine whether
to ignore the existing cache and write over it. Default is set to FALSE.}

\item{output_dir}{A string specifying the directory where the predictors data
file should be saved (default is "03_outputs/3a_model_outputs").}
}
\value{
A data object (`predictor_data`) containing the generated predictors.
}
\description{
This function creates predictors data based on spatial inputs or loads cached
predictors data if the file already exists. It saves the generated data to a
specified directory for reuse and provides progress updates.
}
\examples{
\dontrun{
predictors <- generate_predictor_data(
  country_code = "KEN",
  country_shape = country_sf,
  pop_raster = pop_raster,
  ur_raster = urban_raster,
  adm2_shape = adm2_sf,
  cell_size = 5000,
  output_dir = "03_outputs/3a_model_outputs"
)
}

}
