% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_current_script_path}
\alias{get_current_script_path}
\title{Get the path to the currently running script}
\usage{
get_current_script_path()
}
\value{
A character string with the full path to the current script,
        or `NULL` if it cannot be determined.
}
\description{
This function attempts to determine the file path of the currently executing
R script. It works across multiple contexts:
- In RStudio: returns the path of the active source editor tab.
- In Rscript: extracts the `--file` argument used in script execution.
- When sourced: uses the internal `ofile` value.
}
\examples{
get_current_script_path()

}
