% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{init}
\alias{init}
\title{Initialize Full Pipeline Script and Model Script}
\usage{
init(
  r_script_name = "full_pipeline.R",
  cpp_script_name = "model.cpp",
  path = here::here(),
  open_r_script = TRUE
)
}
\arguments{
\item{r_script_name}{Character. The name of the R script file to be created.
Defaults to `"full_pipeline.R"`.}

\item{cpp_script_name}{Character. The name of the C++ script file to be
created. Defaults to `"model.cpp"`.}

\item{path}{Character. The directory in which to create the scripts.
Defaults to `"."`.}

\item{open_r_script}{Logical. Whether to open the R script automatically in
RStudio (if available). Defaults to `TRUE`.}
}
\value{
A list of character strings with the file paths of the created
scripts. The function also generates success messages upon completion.
}
\description{
This function creates a full pipeline R script and a model C++ script,
saving them to the appropriate folders within the project directory
structure. The folder structure is created using
`AgePopDenom::create_project_structure()`. The R script is automatically
opened in RStudio after creation.
}
\examples{
# Create the full pipeline script and model script
# init()

}
