% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{process_dhs_data}
\alias{process_dhs_data}
\title{Process DHS Data: Merge RDS Files with Shapefiles and Extract Gamma
Parameters}
\usage{
process_dhs_data(
  rds_dir = here::here("01_data", "1a_survey_data", "raw", "pr_records"),
  shp_dir = here::here("01_data", "1a_survey_data", "raw", "shapefiles"),
  output_path = here::here("01_data", "1a_survey_data", "processed",
    "dhs_pr_records_combined.rds")
)
}
\arguments{
\item{rds_dir}{Character. Path to the directory containing raw RDS files.}

\item{shp_dir}{Character. Path to the directory containing shapefiles.}

\item{output_path}{Character. Path to save the final processed dataset as an
RDS file.}
}
\value{
None. Saves the final combined dataset to the specified output path.
}
\description{
This function processes DHS (Demographic and Health Survey) data by:
1. Reading RDS files and shapefiles for each country.
2. Merging demographic data with geographic information.
3. Cleaning and aggregating the data.
4. Extracting gamma parameters for age-related analysis.
}
\details{
The function loops through RDS files, processes each country's data by
merging demographic information with shapefile data, and computes gamma
parameters for age-related analysis. The progress is tracked and displayed
for each country.

The function also filters out incomplete data (e.g., age values of `98`) and
handles labelled data using the `haven::zap_labels` function.

The final output includes two datasets:
1. Outlier-free data.
2. Aggregated age parameter data.
}
\examples{
\dontrun{
# Not run to avoid lengthy data downloads, processing and modelling
# in examples
process_dhs_data(
  rds_dir = "01_data/1a_survey_data/raw/pr_records",
  shp_dir = "01_data/1a_survey_data/raw/shapefiles/",
  output_path = "01_data/1a_survey_data/processed/dhs_pr_records_combined.rds"
)
}

}
