% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{process_gamma_predictions}
\alias{process_gamma_predictions}
\title{Process Gamma Prediction Results}
\usage{
process_gamma_predictions(gamma_prediction)
}
\arguments{
\item{gamma_prediction}{A list containing `scale_pred` and `shape_pred`
matrices from the gamma prediction model.}
}
\value{
A list containing the following elements:
  - `mean_age_pred`: A vector of mean age predictions.
  - `scale_hat`: A vector of mean scale parameters.
  - `shape_hat`: A vector of mean shape parameters.
}
\description{
This function processes gamma prediction results to calculate the mean age
predictions, scale, and shape parameters efficiently.
}
