% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{run_spatial_model}
\alias{run_spatial_model}
\title{Run or Load a Cached Spatial Model}
\usage{
run_spatial_model(
  country_code,
  age_param_data,
  scale_outcome = "log_scale",
  shape_outcome = "log_shape",
  covariates = "urban",
  cpp_script_name = "02_scripts/model_optim",
  output_dir = "03_outputs/3a_model_outputs",
  ignore_cache = FALSE
)
}
\arguments{
\item{country_code}{A string representing the country code (e.g., "KEN").}

\item{age_param_data}{A data frame or tibble containing the data for the model.}

\item{scale_outcome}{A string specifying the outcome variable for the scale
parameter (default is "log_scale").}

\item{shape_outcome}{A string specifying the outcome variable for the shape
parameter (default is "log_shape").}

\item{covariates}{A string or vector of strings representing the covariates
to include in the model (default is "urban").}

\item{cpp_script_name}{A string specifying the path to the CPP script used
for optimization (default is "02_scripts/model_optim").}

\item{output_dir}{A string specifying the directory where the model results
should be saved (default is "03_outputs/3a_model_outputs").}

\item{ignore_cache}{A boolean input which is set to determine whether
to ignore the existing cache and write over it. Default is set to FALSE.}
}
\value{
The fitted spatial model object (`spat_model_param`).
}
\description{
This function fits a spatial model for age parameters or loads a cached
result if the model output file already exists. It saves the fitted model
to a specified directory and provides progress updates.
}
\examples{
\dontrun{
result <- run_spatial_model(
  country_code = "KEN",
  age_param_data = my_data,
  scale_outcome = "log_scale",
  shape_outcome = "log_shape",
  covariates = "urban",
  cpp_script_name = "02_scripts/model_optim",
  output_dir = "03_outputs/3a_model_outputs"
)
}
}
