% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{generate_age_pop_table}
\alias{generate_age_pop_table}
\title{Generate Age Population Data Across Intervals}
\usage{
generate_age_pop_table(
  predictor_data,
  scale_pred,
  shape_pred,
  age_range = c(0, 100),
  interval = 5,
  country_code,
  ignore_cache = FALSE,
  output_dir,
  n_cores = parallel::detectCores() - 2
)
}
\arguments{
\item{predictor_data}{A data frame containing columns: `country`, `region`,
`district`, and `pop`.}

\item{scale_pred}{A matrix of scale predictions (rows correspond to data
points, columns to simulations).}

\item{shape_pred}{A matrix of shape predictions (rows correspond to data
points, columns to simulations).}

\item{age_range}{A numeric vector of length 2 specifying the minimum and
maximum ages for the intervals (e.g., `c(0, 100)`).}

\item{interval}{A numeric value specifying the size of each age interval.}

\item{country_code}{A string representing the lowercase country code,
used for naming the output file.}

\item{ignore_cache}{A boolean input which is set to determine whether
to ignore the existing cache and write over it. Default is set to FALSE.}

\item{output_dir}{A string specifying the directory where the output file
should be saved.}

\item{n_cores}{An integer specifying the number of cores for parallel
processing (default is `parallel::detectCores() - 2`).}
}
\value{
A data frame with population estimates and quantiles across age
  intervals.
}
\description{
This function computes population estimates for different age intervals
based on gamma distributions, aggregates results across simulations,
and returns a summarized data frame with mean, lower, and upper quantiles.
}
\examples{
# final_data <- generate_age_pop_table(
#  predictor_data = predictor_data,
#  scale_pred = scale_pred,
#  shape_pred = shape_pred,
#  age_range = c(0, 100),
#  interval = 5,
#  country_code = "ken",
#  output_dir = "03_outputs/3a_model_outputs"
# )

}
