% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rasterize_data}
\alias{rasterize_data}
\title{Rasterize Spatial Data}
\usage{
rasterize_data(x_coords, y_coords, values, resolution, crs, fun = mean)
}
\arguments{
\item{x_coords}{Numeric vector of x-coordinates (e.g., longitude).}

\item{y_coords}{Numeric vector of y-coordinates (e.g., latitude).}

\item{values}{Numeric vector of values associated with each point.}

\item{resolution}{Numeric, the resolution of the raster in map units
(e.g., meters for EPSG:3857).}

\item{crs}{Character, the coordinate reference system in EPSG format
(e.g., "EPSG:3857").}

\item{fun}{Function to aggregate values in cells (default is `mean`).}
}
\value{
A `terra::SpatRaster` object.
}
\description{
This function converts spatial data with x, y coordinates and a value field
into a raster using a specified resolution and CRS.
}
\examples{
# rast <- rasterize_data(
#   predictor_data$web_x, predictor_data$web_y, pred_list$shape_hat,
#   resolution = 5000, crs = "EPSG:3857"
# )
}
