% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{run_models_with_logging}
\alias{run_models_with_logging}
\title{Run Models for Multiple Countries and Log Results}
\usage{
run_models_with_logging(
  country_codes,
  dhs_data_path = "01_data/1a_dhs_data/processed/dhs_pr_records_combined.rds",
  shape_path = "01_data/1c_shapefiles/district_shape.gpkg",
  pop_raster_path = "01_data/1b_rasters/worldpop_100m",
  ur_raster_path = "01_data/1b_rasters/urban_extent/afurextent.asc",
  log_path = "03_outputs/3a_model_outputs/modelling_log.rds",
  ignore_cache = FALSE
)
}
\arguments{
\item{country_codes}{A character vector of country codes to process.}

\item{dhs_data_path}{Character. Path to DHS data. Default:
"01_data/1a_dhs_data/processed/dhs_pr_records_combined.rds".}

\item{shape_path}{Character. Path to DHS data. Default:
"01_data/1c_shapefiles/district_shape.gpkg".}

\item{pop_raster_path}{Character. Path to DHS data. Default:
"01_data/1b_rasters/worldpop_100m".}

\item{ur_raster_path}{Character. Path to urban-rural extent data. Default:
"01_data/1b_rasters/urban_extent/afurextent.asc".}

\item{log_path}{A character string specifying the file path for the log file.}

\item{ignore_cache}{A boolean input which is set to determine whether
to ignore the existing cache and write over it. Default is set to FALSE.}
}
\value{
None. The function updates the log file at `log_path`.
}
\description{
Runs a model for each country in the provided list, logs the status,
duration, and any errors, and merges with an existing log before saving.
}
\examples{
# run_models_with_logging(
#  country_codes = c("KEN", "UGA"),
#  log_path = "03_outputs/3a_model_outputs/modelling_log.rds"
# )
}
